+++ internal/updater/writer_config_test.go
